+++ CHANGELOG.rst
@@ -1,7 +1,10 @@
 4.1.4 (unreleased)
 ------------------
 
-- Fix error on invalid CORS check
+- Implement default PUT method to be able to replace content
+  [vangheem]
+
+- Fix error on invalid CORS ch
   [vangheem]
 
 - Option to disable inheritance on role permission relation